Move factories, move builders to package
